[Socket] Move remaining icalls and struct from Socket_2_1.cs to Socket.cs
