--- mcs/class/System/System.Net.Sockets/Socket_2_1.cs
@@ -1,68 +1,0 @@
-// System.Net.Sockets.Socket.cs
-//
-// Authors:
-//	Phillip Pearson (pp@myelin.co.nz)
-//	Dick Porter <dick@ximian.com>
-//	Gonzalo Paniagua Javier (gonzalo@ximian.com)
-//	Sridhar Kulkarni (sridharkulkarni@gmail.com)
-//	Brian Nickel (brian.nickel@gmail.com)
-//
-// Copyright (C) 2001, 2002 Phillip Pearson and Ximian, Inc.
-//    http://www.myelin.co.nz
-// (c) 2004-2011 Novell, Inc. (http://www.novell.com)
-//
-
-//
-// Permission is hereby granted, free of charge, to any person obtaining
-// a copy of this software and associated documentation files (the
-// "Software"), to deal in the Software without restriction, including
-// without limitation the rights to use, copy, modify, merge, publish,
-// distribute, sublicense, and/or sell copies of the Software, and to
-// permit persons to whom the Software is furnished to do so, subject to
-// the following conditions:
-// 
-// The above copyright notice and this permission notice shall be
-// included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
-// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
-// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-//
-
-using System;
-using System.Net;
-using System.Collections;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-using System.Threading;
-using System.IO;
-using System.Security;
-using System.Text;
-
-#if !NET_2_1
-using System.Net.Configuration;
-using System.Net.NetworkInformation;
-#endif
-
-namespace System.Net.Sockets {
-
-	public partial class Socket : IDisposable {
-		[StructLayout (LayoutKind.Sequential)]
-		struct WSABUF {
-			public int len;
-			public IntPtr buf;
-		}
-
-		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal static extern void cancel_blocking_socket_operation (Thread thread);
-
-		[MethodImplAttribute(MethodImplOptions.InternalCall)]
-		internal static extern void socket_pool_queue (SocketAsyncCallback d, SocketAsyncResult r);
-	}
-}
-
